wayland: pointer.scale: convert to float

+++ wayland.c
@@ -1733,7 +1733,7 @@ wayl_win_destroy(struct wl_window *win)
 }
 
 bool
-wayl_reload_xcursor_theme(struct seat *seat, int new_scale)
+wayl_reload_xcursor_theme(struct seat *seat, float new_scale)
 {
     if (seat->pointer.theme != NULL && seat->pointer.scale == new_scale) {
         /* We already have a theme loaded, and the scale hasn't changed */
@@ -1766,7 +1766,7 @@ wayl_reload_xcursor_theme(struct seat *seat, int new_scale)
 
     const char *xcursor_theme = getenv("XCURSOR_THEME");
 
-    LOG_INFO("cursor theme: %s, size: %d, scale: %d",
+    LOG_INFO("cursor theme: %s, size: %d, scale: %.2f",
              xcursor_theme ? xcursor_theme : "(null)",
              xcursor_size, new_scale);
